requirements.txt
selenium
beautifulsoup4
requests 
mysql-connector-python==8.0.32
